+++ axum/CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **fixed:** Allow unreachable code in `#[debug_handler]` ([#2014])
 - **change:** Update tokio-tungstenite to 0.19 ([#2021])
 - **change:** axum's MSRV is now 1.63 ([#2021])
+- **added:** Implement `Handler` for `T: IntoResponse`
 
 [#2021]: https://github.com/tokio-rs/axum/pull/2021
 [#2014]: https://github.com/tokio-rs/axum/pull/2014